# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Word API",
    "description": "API for managing words",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ],
  "paths": {
    "/words": {
      "get": {
        "summary": "Get all words",
        "responses": {
          "200": {
            "description": "A list of words",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Word"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new word",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Word"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The newly created word",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Word"
                }
              }
            }
          }
        }
      }
    },
    "/words/{id}": {
      "get": {
        "summary": "Get word by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The word",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Word"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update word by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Word"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated word",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Word"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete word by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Word deleted successfully"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Word": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "word": {
            "type": "string"
          }
        }
      }
    }
  }
}